implemented setuptools and renamed libs to mu
