+++ mu/__main__.py
@@ -1,0 +1,4 @@
+from mu.main import main
+
+if __name__ == '__main__':
+    raise SystemExit(main())
