+++ setup.cfg
@@ -1,0 +1,33 @@
+[metadata]
+name = mu
+version = 1.2.0
+description = Mikrotik Upgrade command line tool
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/spidermila/mikrotik_upgrade
+author = Spidermila
+author_email = spidermila@gmail.com
+license = MIT
+license_files = LICENSE
+classifiers =
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: Implementation :: CPython
+
+[options]
+packages = find:
+install_requires =
+    PyYAML>=6.0.2
+    paramiko>=3.5.0
+    scp>=0.15.0
+python_requires = >=3.11
+
+[options.packages.find]
+exclude =
+    routeros*
+    tests*
+
+[options.entry_points]
+console_scripts =
+    mu = mu.main:main
